refactor: HttpSettings, HttpConnection location update: replace timeout::Settings with TimeoutConfig add: HttpConnector to aws_types::Config and Builder refactor: divide default_provider.rs modules into separate files
