+++ aws/rust-runtime/aws-config/src/test_case.rs
@@ -3,13 +3,17 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 
-use crate::provider_config::{HttpSettings, ProviderConfig};
+use crate::provider_config::ProviderConfig;
+
 use aws_smithy_async::rt::sleep::{AsyncSleep, Sleep, TokioSleep};
 use aws_smithy_client::dvr::{NetworkTraffic, RecordingConnection, ReplayingConnection};
 use aws_smithy_client::erase::DynConnector;
+use aws_smithy_client::http_connector::HttpSettings;
 use aws_types::credentials::{self, ProvideCredentials};
 use aws_types::os_shim_internal::{Env, Fs};
+
 use serde::Deserialize;
+
 use std::collections::HashMap;
 use std::error::Error;
 use std::fmt::Debug;
@@ -178,7 +182,8 @@ impl TestEnvironment {
         let settings = HttpSettings::default();
         let (_test_connector, config) = self.provider_config().await;
         let live_connector =
-            crate::connector::default_connector(&settings, config.sleep()).unwrap();
+            aws_smithy_client::http_connector::default_connector(&settings, config.sleep())
+                .unwrap();
         let live_connector = RecordingConnection::new(live_connector);
         let config = config.with_http_connector(DynConnector::new(live_connector.clone()));
         let provider = make_provider(config).await;
